Add geocoder service url for emdat and ifrc.
  - Remove base.py file from transform and moved to emdat transform
    file.
