+++ .github/workflows/ci.yml
@@ -25,6 +25,7 @@ env:
   EMDAT_AUTHORIZATION_KEY: dummy-value
   IDMC_CLIENT_ID: dummy-value
   IFRC_DATA_URL: dummy-url
+  GEOCODER_URL: dummy-url
   EOAPI_DOMAIN: https://montandon-eoapi.dummy.com
   DESINVENTAR_DATA_URL: https://desinventar.dummy.com
   GFD_CREDENTIAL: dummy-value
