+++ helm/templates/config/secret.yaml
@@ -21,6 +21,7 @@ stringData:
   GFD_CREDENTIAL: {{ required "secrets.GFD_CREDENTIAL" .Values.secrets.GFD_CREDENTIAL | quote }}
   GFD_SERVICE_ACCOUNT: {{ required "secrets.GFD_SERVICE_ACCOUNT" .Values.secrets.GFD_SERVICE_ACCOUNT | quote }}
   IFRC_DATA_URL: {{ required "secrets.IFRC_DATA_URL" .Values.secrets.IFRC_DATA_URL | quote }}
+  
 
   # Database
   {{- if .Values.postgresql.enabled }}
