+++ helm/values.yaml
@@ -193,6 +193,7 @@ secrets:
   GFD_CREDENDIAL:
   GFD_SERVICE_ACCOUNT:
   IFRC_DATA_URL:
+  GEOCODER_URL:
 # NOTE: Used to pass additional secrets to api/worker containers
 # NOTE: Not used by azure vault
 secretsAdditional:
