+++ main/settings.py
@@ -77,9 +77,12 @@ env = environ.Env(
     DESINVENTAR_DATA_URL=str,
     # ETL Load configs
     EOAPI_DOMAIN=str,  # http://montandon-eoapi.ifrc.org
+    GEOCODER_URL=str,
 )
 
 DESINVENTAR_DATA_URL = env("DESINVENTAR_DATA_URL")
+
+GEOCODER_URL = env("GEOCODER_URL")
 
 GFD_SERVICE_ACCOUNT = env("GFD_SERVICE_ACCOUNT")
